+++ homie-state/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @siper/homie-state
 
+## 1.1.5
+
+### Patch Changes
+
+- server snapshot
+
+## 1.1.4
+
+### Patch Changes
+
+- add server snapshot
+
+## 1.1.3
+
+### Patch Changes
+
+- q
+
+## 1.1.2
+
+### Patch Changes
+
+- version up
+
+## 1.1.1
+
+### Patch Changes
+
+- version up
+
+## 1.1.0
+
+### Minor Changes
+
+- homie state
+
 ## 1.0.2
 
 ### Patch Changes